% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simX.R
\docType{data}
\name{simX}
\alias{simX}
\title{Simulated Data}
\format{
A four-way array having \code{p = 2} variables in the rows, \code{r = 3} variables in the columns,
\code{n = 50} data observations and \code{t = 10} times.
}
\usage{
data(simX)
}
\description{
A simulated four-way dataset with 2 states and VII - EE covariance structure.
}
\keyword{datasets}
